# docs/admin.openapi.yaml
...
openapi: 3.0.3
info:
  title: MetroMatrix admin console API
  version: 2.0.0
  description: >-
    Every admin endpoint (/api/admin, /api/v1/admin, /api/shopping/admin) answers in one envelope:


    - success: `{ success: true, data, meta? }` — `meta` pages lists (`page, limit, total, pages, nextCursor`); `limit`
    is clamped to 1–100.

    - failure: `{ success: false, error: { code, message, details? }, requestId }` — clients branch on `error.code` (see
    src/utils/errorCodes.js).


    Authentication: `Authorization: Bearer <accessToken>` from POST /api/admin/auth/login. Access tokens are
    short-lived; renew with POST /api/admin/auth/refresh-token (rotates the refresh token — replaying an old one revokes
    the session).

    A session may be restricted (`restrict: password_change | totp_enrol`) until the admin fixes their account.


    `x-guards` lists the permission guard of each operation (source: docs/ROUTES.json). Money, personal data and audit
    reads are guarded too.

    Status lists, labels, tones, limits and reference data come from GET /api/admin/meta — never hardcode them in a
    client.
servers:
  - url: /
    description: API host (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationFailed:
      description: Validation failed (details.fields lists the problems)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Not signed in, token expired, or session revoked/idle
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Missing permission, super admin required, or session restricted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflicts with the current state (e.g. DELETE_BLOCKED, INVALID_TRANSITION)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyAttempts:
      description: Too many failed attempts — see Retry-After
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        requestId:
          type: string
          nullable: true
    Permissions:
      type: object
      additionalProperties:
        type: boolean
    AdminProfile:
      type: object
      required:
        - id
        - email
        - fullName
        - role
        - isSuperAdmin
        - permissions
        - isActive
        - mustChangePassword
        - twoFactorEnabled
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum:
            - super_admin
            - admin
            - moderator
        isSuperAdmin:
          type: boolean
        permissions:
          $ref: '#/components/schemas/Permissions'
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        passwordChangedAt:
          type: string
          format: date-time
          nullable: true
        twoFactorEnabled:
          type: boolean
        lastLoginDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        restrict:
          type: string
          nullable: true
          enum:
            - password_change
            - totp_enrol
            - null
        liveSessions:
          type: integer
...
    LoginResult:
      type: object
      description: >-
        `step: 'signed_in'` carries the session tokens; `step: 'totp_required'` carries a 5-minute challenge token for
        POST /auth/login/totp.
      required:
        - step
      properties:
        step:
          type: string
          enum:
            - signed_in
            - totp_required
        challengeToken:
          type: string
        expiresInSeconds:
          type: integer
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        sessionId:
          type: string
        restrict:
          type: string
          nullable: true
        admin:
          $ref: '#/components/schemas/AdminProfile'
...
paths:
...
  /api/admin/auth/login:
    post:
      tags:
        - Authentication
      operationId: postAdminAuthLogin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/LoginResult'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyAttempts'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                deviceLabel:
                  type: string
      description: >-
        401: wrong email/password or TOTP code (identical for unknown emails). 403: account deactivated (only after the
        password is proven). 429: locked out — see Retry-After.
...