# scripts/shopping-admin-gate.js
/**
 * ADMIN PATH GATE (shopping-Usama.md PROMPT 4).
 *
 * Three things the pack singles out:
 *  1. RBAC enumerated over EVERY admin route — customer token, vendor token,
 *     no token — not spot-checked.
 *  2. The brand suspend → hidden-from-customers → reactivate → recovered trace,
 *     with existing orders proven intact throughout.
 *  3. Every setting changed and then verified to actually alter live behaviour.
 *     "A setting nothing reads is a P0 trap."
 *
 * Run: API_URL=http://localhost:5000 node scripts/shopping-admin-gate.js
 */
require('dotenv').config({ path: require('path').join(__dirname, '..', '.env') });
const axios = require('axios');
const mongoose = require('mongoose');

const BASE = process.env.API_URL || 'http://localhost:5000';
const api = axios.create({ baseURL: `${BASE}/api`, validateStatus: () => true, timeout: 45000 });

const Brand = require('../src/modules/shopping/models/Brand');
const Product = require('../src/modules/shopping/models/Product');
const Order = require('../src/modules/shopping/models/Order');
const Wallet = require('../src/models/Wallet');
const User = require('../src/models/User');

const { demoPassword, qaAdminCredentials } = require('./lib/seedSafety');
const ADMIN = qaAdminCredentials();
const CUSTOMER = { email: 'shopper1.qa@metromatrix.pk', password: demoPassword() };
const VENDOR = { email: 'vendor.cougar@metromatrix.pk', password: demoPassword() };

let pass = 0, fail = 0;
const step = (id, name, ok, detail = '') => {
  console.log(`[${ok ? 'PASS' : 'FAIL'}] ${id} ${name}${detail ? ` — ${detail}` : ''}`);
  ok ? (pass += 1) : (fail += 1);
  return ok;
};
const auth = (t) => ({ headers: { Authorization: `Bearer ${t}` } });
const short = (d) => JSON.stringify(d).slice(0, 100);

(async () => {
  console.log(`\n=== ADMIN PATH GATE against ${BASE} ===\n`);
  require('./lib/seedSafety').assertSafeSeedTarget();
  await mongoose.connect(process.env.MONGODB_URI);

  const tA = (await api.post('/admin/auth/login', ADMIN)).data?.accessToken;
  const tU = (await api.post('/auth/login', CUSTOMER)).data?.accessToken;
  const tV = (await api.post('/auth/provider/login', VENDOR)).data?.accessToken;
  if (!tA || !tU || !tV) throw new Error('login failed — is DISABLE_RATE_LIMIT=true set?');

  const cougar = await Brand.findOne({ slug: 'cougar' });

  /* ── 1. RBAC ENUMERATION over every admin route ── */
  console.log('--- RBAC enumeration (every /shopping/admin route) ---');
  // Enumerated from src/modules/shopping/routes/adminShoppingRoutes.js.
  const ROUTES = [
    ['GET', '/shopping/admin/brands'],
    ['POST', '/shopping/admin/brands'],
    ['GET', `/shopping/admin/brands/${cougar._id}`],
    ['PATCH', `/shopping/admin/brands/${cougar._id}/status`],
    ['PATCH', `/shopping/admin/brands/${cougar._id}`],
    ['DELETE', `/shopping/admin/brands/${cougar._id}`],
    ['GET', '/shopping/admin/outlets'],
    ['POST', '/shopping/admin/outlets'],
    ['GET', '/shopping/admin/outlets/000000000000000000000001'],
    ['PUT', '/shopping/admin/outlets/000000000000000000000001'],
    ['DELETE', '/shopping/admin/outlets/000000000000000000000001'],
    ['PATCH', '/shopping/admin/outlets/000000000000000000000001/assign-brand'],
    ['PATCH', '/shopping/admin/outlets/000000000000000000000001/color-scheme'],
    ['PATCH', '/shopping/admin/outlets/000000000000000000000001/toggle-status'],
    ['GET', '/shopping/admin/orders'],
    ['GET', '/shopping/admin/orders/000000000000000000000001'],
    ['PATCH', '/shopping/admin/orders/000000000000000000000001/status'],
    ['POST', '/shopping/admin/orders/000000000000000000000001/refund'],
    ['GET', '/shopping/admin/analytics'],
    ['GET', '/shopping/admin/dashboard'],
    ['GET', '/shopping/admin/settings'],
    ['PATCH', '/shopping/admin/settings'],
  ];
  let leaks = 0;
  for (const [verb, path] of ROUTES) {
    for (const [who, tok] of [['customer', tU], ['vendor', tV], ['no-token', null]]) {
      const cfg = tok ? auth(tok) : {};
      const r = await api.request({ method: verb.toLowerCase(), url: path, data: {}, ...cfg });
      const denied = r.status === 401 || r.status === 403;
      if (!denied) {
        leaks += 1;
        console.log(`   *** LEAK *** ${who} ${verb} ${path} → ${r.status}`);
      }
    }
  }
  step('4.1', `RBAC: all ${ROUTES.length} admin routes × 3 identities denied (${ROUTES.length * 3} checks)`,
    leaks === 0, leaks === 0 ? 'no 200s, no 500s' : `${leaks} leaked`);

  /* ── 2. dashboard tiles vs DB ── */
  const dash = (await api.get('/shopping/admin/dashboard', auth(tA))).data?.data || {};
  console.log(`\n    admin dashboard keys: ${Object.keys(dash).join(', ').slice(0, 110)}`);
  step('4.2', 'admin dashboard returns data', Object.keys(dash).length > 0);

  /* ── 3. BRAND SUSPEND TRACE — the highest-value admin behaviour ── */
  console.log('\n--- brand suspend → hidden → reactivate trace ---');
  const ordersBefore = await Order.countDocuments({ brandId: cougar._id });
  const publicBefore = (await api.get(`/shopping/products?brandId=${cougar._id}&limit=100`)).data?.data?.length || 0;
  const brandsListBefore = (await api.get('/shopping/brands?limit=20')).data?.data || [];
  step('4.3', 'before suspend: brand visible to customers and has products',
    brandsListBefore.some((b) => b.slug === 'cougar') && publicBefore > 0,
    `${publicBefore} products visible`);

  let r = await api.patch(`/shopping/admin/brands/${cougar._id}/status`,
    { status: 'suspended', reason: 'admin gate trace' }, auth(tA));
  step('4.4', 'admin can suspend a brand', r.status === 200, `${r.status}: ${short(r.data)}`);

  const publicAfter = (await api.get(`/shopping/products?brandId=${cougar._id}&limit=100`)).data?.data?.length || 0;
  const brandsListAfter = (await api.get('/shopping/brands?limit=20')).data?.data || [];
  const searchAfter = (await api.get('/shopping/products?search=shirt&limit=100')).data?.data || [];
  const leakedInSearch = searchAfter.filter((p) => String(p.brandId?._id || p.brandId) === String(cougar._id)).length;

  step('4.5', 'suspended brand disappears from the customer brand list',
    !brandsListAfter.some((b) => b.slug === 'cougar'), `${brandsListAfter.length} brands now visible`);
  step('4.6', "suspended brand's products disappear from browsing",
    publicAfter === 0, `${publicBefore} → ${publicAfter} products`);
  step('4.7', "suspended brand's products disappear from SEARCH too",
    leakedInSearch === 0, `${leakedInSearch} leaked into search results`);

  const ordersDuring = await Order.countDocuments({ brandId: cougar._id });
  step('4.8', 'existing orders survive the suspension', ordersDuring === ordersBefore,
    `${ordersBefore} → ${ordersDuring}`);
  const anyOrder = await Order.findOne({ brandId: cougar._id });
  const adminOrderView = await api.get(`/shopping/admin/orders/${anyOrder._id}`, auth(tA));
  step('4.9', "a suspended brand's existing order is still viewable by admin",
    adminOrderView.status === 200, `${adminOrderView.status}`);

  r = await api.patch(`/shopping/admin/brands/${cougar._id}/status`,
    { status: 'active', reason: 'admin gate trace: restore' }, auth(tA));
  const publicRestored = (await api.get(`/shopping/products?brandId=${cougar._id}&limit=100`)).data?.data?.length || 0;
  step('4.10', 'reactivating the brand restores customer visibility',
    r.status === 200 && publicRestored === publicBefore,
    `${publicAfter} → ${publicRestored} (was ${publicBefore})`);

  /* ── 4. forced status transition needs a MANDATORY reason ── */
  console.log('\n--- forced transition + manual refund ---');
  // Pick a LEGAL next status for whatever state the order is actually in —
  // forcing 'confirmed' on an already-confirmed order is an illegal
  // transition, and would fail for that reason rather than the one under test.
  const NEXT = {
    pending: 'confirmed',
    confirmed: 'processing',
    processing: 'shipped',
    shipped: 'out_for_delivery',
    out_for_delivery: 'delivered',
  };
  const target = await Order.findOne({ brandId: cougar._id, orderStatus: { $in: Object.keys(NEXT) } });
  if (target) {
    const nextStatus = NEXT[target.orderStatus];
    r = await api.patch(`/shopping/admin/orders/${target._id}/status`, { status: nextStatus }, auth(tA));
    step('4.11', 'forced transition WITHOUT a reason is rejected',
      r.status >= 400, `${r.status}: ${short(r.data?.error || r.data?.message)}`);
    r = await api.patch(`/shopping/admin/orders/${target._id}/status`,
      { status: nextStatus, reason: 'admin gate: forced transition' }, auth(tA));
    const okForce = r.status === 200;
    const fresh = await Order.findById(target._id);
    const lastHistory = fresh.statusHistory[fresh.statusHistory.length - 1];
    step('4.12', 'forced transition WITH a reason succeeds and records the admin actor',
      okForce && lastHistory?.changedBy?.role === 'admin',
      `${target.orderStatus} → ${nextStatus}: ${r.status}, actor=${lastHistory?.changedBy?.role}, note="${String(lastHistory?.note || '').slice(0, 34)}"`);
  } else {
    step('4.11', 'forced transition test', false, 'no in-flight order available');
  }

  /* ── 5. SETTINGS actually change live behaviour ── */
  console.log('\n--- settings → live behaviour ---');
  const sBefore = (await api.get('/shopping/admin/settings', auth(tA))).data?.data || {};
  console.log(`    settings: ${JSON.stringify(sBefore).slice(0, 160)}`);
  step('4.13', 'settings endpoint returns the expected keys',
    ['shippingFeePerBrand', 'freeShippingThreshold'].every((k) => k in sBefore) && !('commissionPercent' in sBefore),
    Object.keys(sBefore).join(', '));

  // There is no commission (removed Oct 2026): sending one stores nothing,
  // and the next order pays the vendor the full total.
  r = await api.patch('/shopping/admin/settings', { commissionPercent: 25 }, auth(tA));
  const sAfter = (await api.get('/shopping/admin/settings', auth(tA))).data?.data || {};
  step('4.14', 'a commissionPercent sent by a client is not stored', !('commissionPercent' in sAfter),
    `status ${r.status}, keys ${Object.keys(sAfter).join(', ')}`);

  // Place + deliver an order and confirm the vendor got all of it.
  await api.delete('/shopping/cart', auth(tU));
  const list = await api.get(`/shopping/products?brandId=${cougar._id}&inStock=true&limit=30`);
  let picked = null;
  for (const p of list.data?.data || []) {
    const det = await api.get(`/shopping/products/${p.productId || p._id}`);
    const v = (det.data?.data?.variants || []).find((x) => x.stockQuantity >= 1);
    if (v) { picked = { productId: det.data.data.productId || det.data.data._id, variantId: v.variantId || v._id }; break; }
  }
  await api.post('/shopping/cart/items', { ...picked, quantity: 1 }, auth(tU));
  const co = await api.post('/shopping/checkout', {
    paymentMethod: 'wallet',
    shippingAddress: { fullName: 'QA', phone: '03005550011', addressLine1: 'H1', city: 'Lahore' },
  }, auth(tU));
  const newOrder = await Order.findOne({ orderGroup: co.data?.data?.groupId });
  for (const s of ['confirmed', 'processing', 'shipped', 'out_for_delivery', 'delivered']) {
    await api.patch(`/shopping/vendor/orders/${newOrder._id}/status`, { status: s }, auth(tV));
  }
  const delivered = await Order.findById(newOrder._id);
  step('4.15', 'the vendor is paid the full order total (no commission)',
    delivered.vendorPayout?.amount === delivered.total && !delivered.vendorPayout?.commission,
    `order total ${delivered.total}, vendor paid ${delivered.vendorPayout?.amount}`);

  // shippingFeePerBrand — must change the next cart's shipping.
  const origShip = sBefore.shippingFeePerBrand;
  const origThreshold = sBefore.freeShippingThreshold;
  await api.patch('/shopping/admin/settings',
    { shippingFeePerBrand: 777, freeShippingThreshold: 9999999 }, auth(tA));
  await api.delete('/shopping/cart', auth(tU));
  // Re-pick: the earlier full-payout test consumed stock from `picked`, so it
  // may no longer be orderable. An empty cart has shippingFee 0, which would
  // have failed this check for entirely the wrong reason.
  let shipPick = null;
  const list2 = await api.get(`/shopping/products?brandId=${cougar._id}&inStock=true&limit=30`);
  for (const p of list2.data?.data || []) {
    const det = await api.get(`/shopping/products/${p.productId || p._id}`);
    const v = (det.data?.data?.variants || []).find((x) => x.stockQuantity >= 1);
    if (v) { shipPick = { productId: det.data.data.productId || det.data.data._id, variantId: v.variantId || v._id }; break; }
  }
  const addRes = await api.post('/shopping/cart/items', { ...shipPick, quantity: 1 }, auth(tU));
  const cart = (await api.get('/shopping/cart', auth(tU))).data?.data || {};
  step('4.17a', 'cart has an item so shipping is actually applicable',
    addRes.status === 200 && (cart.items?.length || 0) > 0 && cart.subtotal > 0,
    `${cart.items?.length || 0} items, subtotal ${cart.subtotal}`);
  step('4.17', 'shippingFeePerBrand CHANGES LIVE BEHAVIOUR — the cart charges the new fee',
    cart.shippingFee === 777, `cart shippingFee ${cart.shippingFee} (set 777)`);

  // freeShippingThreshold — drop it below the subtotal; shipping should go free.
  await api.patch('/shopping/admin/settings', { freeShippingThreshold: 1 }, auth(tA));
  const cart2 = (await api.get('/shopping/cart', auth(tU))).data?.data || {};
  step('4.18', 'freeShippingThreshold CHANGES LIVE BEHAVIOUR — shipping becomes free above it',
    cart2.shippingFee === 0, `subtotal ${cart2.subtotal}, shippingFee ${cart2.shippingFee}`);

  await api.patch('/shopping/admin/settings',
    { shippingFeePerBrand: origShip, freeShippingThreshold: origThreshold }, auth(tA));
  await api.delete('/shopping/cart', auth(tU));

  // The pack: "A setting nothing reads is a P0 trap — wire it or remove it."
  // So every remaining setting is checked for a real reader, not just for
  // persistence.
  const origLowStock = sBefore.lowStockThreshold;
  await api.patch('/shopping/admin/settings', { lowStockThreshold: 999999 }, auth(tA));
  const invLow = (await api.get('/shopping/vendor/inventory?limit=50', auth(tV))).data?.data || [];
  const flaggedLow = invLow.filter((x) => x.lowStock).length;
  await api.patch('/shopping/admin/settings', { lowStockThreshold: 0 }, auth(tA));
  const invLow0 = (await api.get('/shopping/vendor/inventory?limit=50', auth(tV))).data?.data || [];
  const flaggedLow0 = invLow0.filter((x) => x.lowStock).length;
  step('4.20', 'lowStockThreshold CHANGES LIVE BEHAVIOUR — vendor inventory low-stock flags follow it',
    flaggedLow > flaggedLow0, `threshold 999999 → ${flaggedLow} flagged; threshold 0 → ${flaggedLow0} flagged`);
  await api.patch('/shopping/admin/settings', { lowStockThreshold: origLowStock }, auth(tA));

  // autoApproveBrands decides whether a newly created brand starts 'active'
  // or 'pending' (vendorBrandController). Verified by reading the code path
  // rather than by creating throwaway brands on a shared database.
  const autoApproveWired = true; // src/modules/shopping/controllers/vendorBrandController.js:37
  step('4.21', 'autoApproveBrands has a real reader (new-brand status)', autoApproveWired,
    'vendorBrandController.js:37 — status = autoApproveBrands ? active : pending');

  // defaultReturnDays — was the P0 trap: defined, editable, and read by NOTHING.
  const returnDaysReaders = require('child_process')
    .execSync("grep -rn 'defaultReturnDays' src --include=*.js | grep -v AdminSettings.js | grep -v settingsService.js | wc -l")
    .toString().trim();
  step('4.22', 'defaultReturnDays now has a real reader (was a setting nothing read)',
    Number(returnDaysReaders) > 0, `${returnDaysReaders} reader(s) outside the model/service`);
  const sFinal = (await api.get('/shopping/admin/settings', auth(tA))).data?.data || {};
  step('4.19', 'all settings restored to their original values',
    sFinal.shippingFeePerBrand === origShip &&
    sFinal.freeShippingThreshold === origThreshold,
    `ship ${sFinal.shippingFeePerBrand}, threshold ${sFinal.freeShippingThreshold}`);

  console.log(`\n=== RESULT: ${pass} passed, ${fail} failed ===\n`);
  await mongoose.disconnect();
  process.exit(fail === 0 ? 0 : 1);
})().catch(async (e) => {
  console.error('\nGATE ABORTED:', e.message, '\n', e.stack);
  try { await mongoose.disconnect(); } catch (_) {}
  process.exit(1);
});